feat: Introduce test configuration and fixtures for async testing
- Created a new `conftest.py` file to set up testing environment with in-memory SQLite database.
- Added fixtures for database preparation and async HTTP client.
- Refactored `test_users.py` to utilize the new test setup, improving test organization and clarity.
